feat: change pages structure to avoid dashboard menu refresh
