+++ src/app/(dashboard)/jobmonitor/layout.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 export default function JobMonitorLayout({
   children,
 }: {
