--- src/app/jobmonitor/layout.tsx
@@ -1,7 +1,0 @@
-export default function JobMonitorLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <section>{children}</section>;
-}
