+++ src/components/applications/JobMonitor.tsx
@@ -15,19 +15,9 @@ export default function JobMonitor() {
   const theme = useMUITheme();
 
   return (
-    <Dashboard>
-      <MUIThemeProvider theme={theme}>
-        <CssBaseline />
-        <Box
-          sx={{
-            ml: "5%",
-            mr: "5%",
-          }}
-        >
-          <h2>Job Monitor</h2>
-          <JobDataTable />
-        </Box>
-      </MUIThemeProvider>
-    </Dashboard>
+    <div>
+      <h2>Job Monitor</h2>
+      <JobDataTable />
+    </div>
   );
 }
